requirements.txt
alembic==1.1.0
amqp==2.5.1
aniso8601==7.0.0
asn1crypto==0.24.0
Babel==2.7.0
bcrypt==3.1.7
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
configparser==3.8.1
contextlib2==0.5.5
cryptography==2.7
debtcollector==1.22.0
decorator==4.4.0
defusedxml==0.6.0
dnspython==1.16.0
dogpile.cache==0.7.1
enum34==1.1.6
eventlet==0.25.1
extras==1.0.0
fasteners==0.15
fixtures==3.0.0
Flask==1.1.1
Flask-RESTful==0.3.7
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.3.0
futurist==1.9.0
greenlet==0.4.15
idna==2.8
importlib-metadata==0.20
ipaddress==1.0.22
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==2.6.0
keystone==15.0.0
keystoneauth1==3.17.1
keystonemiddleware==7.0.1
kombu==4.6.4
linecache2==1.0.0
Mako==1.1.0
MarkupSafe==1.1.1
monotonic==1.5
more-itertools==5.0.0
msgpack==0.6.1
netaddr==0.7.19
netifaces==0.10.9
oauthlib==3.1.0
os-service-types==1.7.0
oslo.cache==1.37.0
oslo.concurrency==3.30.0
oslo.config==6.11.1
oslo.context==2.23.0
oslo.db==5.0.2
oslo.i18n==3.24.0
oslo.log==3.44.1
oslo.messaging==10.2.0
oslo.middleware==3.38.1
oslo.policy==2.3.1
oslo.serialization==2.29.2
oslo.service==1.40.1
oslo.upgradecheck==0.3.2
oslo.utils==3.41.1
osprofiler==2.8.2
passlib==1.7.1
Paste==3.2.0
PasteDeploy==2.0.1
pathlib2==2.3.4
pbr==5.4.3
prettytable==0.7.2
pycadf==2.10.0
pycparser==2.19
pyinotify==0.9.6
PyJWT==1.7.1
pyOpenSSL==19.0.0
pyparsing==2.4.2
pysaml2==4.8.0
python-dateutil==2.8.0
python-editor==1.0.4
python-keystoneclient==3.20.0
python-mimeparse==1.6.0
pytz==2019.2
PyYAML==5.1.2
repoze.lru==0.7
requests==2.22.0
rfc3986==1.3.2
Routes==2.4.1
scandir==1.10.0
scrypt==0.8.13
six==1.12.0
SQLAlchemy==1.3.8
sqlalchemy-migrate==0.12.0
sqlparse==0.3.0
statsd==3.3.0
stevedore==1.31.0
Tempita==0.5.2
testresources==2.0.1
testscenarios==0.5.0
testtools==2.3.0
traceback2==1.4.0
unittest2==1.1.0
urllib3==1.25.3
vine==1.3.0
WebOb==1.8.5
Werkzeug==0.15.6
wrapt==1.11.2
yappi==1.0
zipp==0.6.0
